Inject a build dependency for the builder class for generic Build.pm support (#2)
For old versions of zef, distros may ship a generic Build.pm that just
loads the appropriate Distribution::Builder and runs that. But we need
to ensure that it's actually installed, so we inject it into the
build-depends.
New zef versions will read the meta-version 1 data and know about the
"builder" key right away.