use a cpan rsync mirror if main one is down
like it is right now. this is the same mirror i put into modules.perl6.org,
so that both would fail consistently, and noticing one is broken lets us
immediately know to fix the other one as well.

Could potentially grab the mirrors list as json and go through all rsync
mirrors, but for now this should be good enough.